requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
pyyaml>=5.4.1
pytest>=6.2.5
aiohttp>=3.8.0
websockets>=10.0
cryptography>=3.4.7

# FastAPI and dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]

# WebSocket support
websockets>=10.0

# Security and OAuth
python-jose[cryptography]
passlib[bcrypt]

# HTTP client
aiohttp>=3.8.0

# Utilities
python-dotenv>=0.19.0
backoff>=2.0.0
